+++ src/components/modal/v2/parts/OfferCard.jsx
@@ -1,7 +1,7 @@
 /** @jsx h */
 import { h } from 'preact';
 
-const OfferCard = ({ offer: { content, meta }, useV4Design }) => {
+const OfferCard = ({ offer: { content, meta }, useV4Design, useNewCheckoutDesign }) => {
     const { termsLabel } = content;
     const aprRemoveTrailingZeros = meta?.apr.replace(/\D00$/, '');
     const aprFieldTitle = aprRemoveTrailingZeros === '0' ? termsLabel?.zeroApr : termsLabel?.nonZeroApr;
@@ -24,7 +24,11 @@ const OfferCard = ({ offer: { content, meta }, useV4Design }) => {
                     <p className="offer__field-value">{meta?.formattedTotalInterest}</p>
                 </div>
                 <div className="offer__field-col">
-                    <strong className={`offer__field-title ${useV4Design === 'true' ? 'v4Design' : ''}`}>
+                    <strong
+                        className={`offer__field-title ${useV4Design === 'true' ? 'v4Design' : ''} ${
+                            useNewCheckoutDesign === 'true' ? 'checkout' : ''
+                        }`}
+                    >
                         {termsLabel?.total}
                     </strong>
                     <strong className="offer__field-value">{meta?.formattedTotalCost}</strong>