+++ src/library/zoid/message/component.js
@@ -88,6 +88,12 @@ export default createGlobalVariableGetter('__paypal_credit_message__', () =>
                 required: false,
                 value: validate.placement
             },
+            pageType: {
+                type: 'string',
+                queryParam: 'page_type',
+                required: false,
+                value: validate.pageType
+            },
             style: {
                 type: 'object',
                 serialization: 'json',
@@ -125,6 +131,12 @@ export default createGlobalVariableGetter('__paypal_credit_message__', () =>
                 queryParam: 'ec_token',
                 required: false,
                 value: validate.ecToken
+            },
+            contextualComponents: {
+                type: 'string',
+                queryParam: 'contextual_components',
+                required: false,
+                value: validate.contextualComponents
             },
             // Callbacks
             onClick: {
@@ -467,7 +479,7 @@ export default createGlobalVariableGetter('__paypal_credit_message__', () =>
                 type: 'string',
                 queryParam: 'features',
                 required: false,
-                value: getFeatures
+                value: validate.features ?? getFeatures
             }
         }
     })