+++ src/utils/sdk.js
@@ -19,8 +19,9 @@ import {
     getStorageID as getSDKStorageID,
     getStorageState as getSDKStorageState,
     getPayPalDomain as getSDKPayPalDomain,
-    getPayPalAPIDomain as getSDKPayPalAPIDomain,
-    getDisableSetCookie as getSDKDisableCookie
+    getDisableSetCookie as getSDKDisableCookie,
+    getPageType as getSDKPageType,
+    getPayPalAPIDomain as getSDKPayPalAPIDomain
 } from '@paypal/sdk-client/src';
 
 import { TAG } from './constants';
@@ -71,7 +72,13 @@ export function getAccount() {
         return undefined;
     }
 }
-
+export function getPageType() {
+    if (__MESSAGES__.__TARGET__ === 'SDK') {
+        return getSDKPageType();
+    } else {
+        return undefined;
+    }
+}
 export function getNonce() {
     if (__MESSAGES__.__TARGET__ === 'SDK') {
         return getCSPNonce();